Revert "[APG-454] refactor"
This reverts commit b7d9ba93d5ae3d3d5e474d5fb5608442af65282f.
